feat(ui): moves add population data to steps dir
