+++ app/src/app/[lng]/onboarding/setup/page.tsx
@@ -50,16 +50,10 @@ import type { TFunction } from "i18next";
 import dynamic from "next/dynamic";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
-import type {
-  Control,
-  FieldErrors,
-  SubmitHandler,
-  UseFormRegister,
-} from "react-hook-form";
-import { Controller, useForm } from "react-hook-form";
+import type { SubmitHandler, UseFormRegister } from "react-hook-form";
+import { useForm } from "react-hook-form";
 import { Trans } from "react-i18next/TransWithoutContext";
 import { MdOutlineAspectRatio, MdOutlinePeopleAlt } from "react-icons/md";
-import FormattedThousandsNumberInput from "@/app/[lng]/onboarding/setup/FormattedThousandsNumberInput";
 
 import {
   CalenderIcon,
@@ -69,6 +63,7 @@ import {
 import { CircleFlag } from "react-circle-flags";
 import SelectCityStep from "../steps/select-city-steps";
 import SetInventoryDetailsStep from "../steps/add-inventory-details-step";
+import SetPopulationDataStep from "../steps/add-population-data-step";
 
 const CityMap = dynamic(() => import("@/components/CityMap"), { ssr: false });
 
@@ -141,385 +136,6 @@ export function CustomRadio(props: CustomRadioProps) {
       >
         {props.isChecked ? <Icon as={InventoryButtonCheckIcon} /> : ""}
         {props.children}
-      </Box>
-    </Box>
-  );
-}
-
-function SetPopulationDataStep({
-  t,
-  register,
-  errors,
-  control,
-  years,
-  ocCityData,
-  watch,
-  setValue,
-  numberOfYearsDisplayed,
-}: {
-  t: TFunction;
-  register: UseFormRegister<Inputs>;
-  errors: FieldErrors<Inputs>;
-  control: Control<Inputs>;
-  years: number[];
-  watch: Function;
-  ocCityData?: OCCityAttributes;
-  setOcCityData: (cityData: OCCityAttributes) => void;
-  setData: (data: OnboardingData) => void;
-  setValue: any;
-  numberOfYearsDisplayed: number;
-}) {
-  const yearInput = watch("year");
-  const year: number | null = yearInput ? parseInt(yearInput) : null;
-
-  const locode = ocCityData?.actor_id;
-  const { data: cityData } = useGetOCCityDataQuery(locode!, {
-    skip: !locode,
-  });
-  const countryLocode =
-    locode && locode.length > 0 ? locode.split(" ")[0] : null;
-  const { data: countryData } = useGetOCCityDataQuery(countryLocode!, {
-    skip: !countryLocode,
-  });
-  const regionLocode = cityData?.is_part_of;
-  const { data: regionData } = useGetOCCityDataQuery(regionLocode!, {
-    skip: !regionLocode,
-  });
-  const [countryPopulationSourceName, setCountryPopulationSourceName] =
-    useState<string>("");
-  useEffect(() => {
-    if (cityData && year) {
-      const population = findClosestYear(
-        cityData.population,
-        year,
-        numberOfYearsDisplayed,
-      );
-      if (!population) {
-        console.error("Failed to find population data for city");
-        return;
-      }
-      setValue("cityPopulation", population.population);
-      setValue("cityPopulationYear", population.year);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [cityData, year, setValue]);
-
-  useEffect(() => {
-    if (regionData && year) {
-      const population = findClosestYear(
-        regionData.population,
-        year,
-        numberOfYearsDisplayed,
-      );
-      if (!population) {
-        console.error("Failed to find population data for region");
-        return;
-      }
-      setValue("regionPopulation", population.population);
-      setValue("regionPopulationYear", population.year);
-    }
-  }, [regionData, year, setValue]);
-
-  useEffect(() => {
-    if (countryData && year) {
-      const population = findClosestYear(
-        countryData.population,
-        year,
-        numberOfYearsDisplayed,
-      );
-      if (!population) {
-        console.error("Failed to find population data for region");
-        return;
-      }
-      let [{ datasource }] = countryData.population;
-      console.log("datasource_name: ", datasource.name);
-      setCountryPopulationSourceName(datasource.name);
-      setValue("countryPopulation", population.population);
-      setValue("countryPopulationYear", population.year);
-      const keys = Object.keys(countryData.emissions);
-      const sourceId = keys.find((id) => id.startsWith("UNFCCC"));
-
-      if (sourceId) {
-        const emissionsData: CountryEmissionsEntry[] =
-          countryData.emissions[sourceId].data;
-        const emissions = emissionsData.find(
-          (e) => e.year === year,
-        )?.total_emissions;
-        if (emissions == null) {
-          console.error("Failed to find country emissions for ", year);
-        }
-        setValue("totalCountryEmissions", emissions);
-      }
-    }
-  }, [countryData, year, setValue]);
-
-  return (
-    <Box w="full">
-      <Box
-        minW={400}
-        w="full"
-        display="flex"
-        flexDir="column"
-        gap="24px"
-        mb="48px"
-      >
-        <Heading size="xl">{t("setup-population-data-heading")}</Heading>
-        <Text
-          color="content.tertiary"
-          fontSize="body.lg"
-          fontStyle="normal"
-          fontWeight="400"
-          letterSpacing="wide"
-        >
-          {t("setup-population-data-description")}
-        </Text>
-      </Box>
-      <Box
-        w="full"
-        py="36px"
-        borderBottomWidth="2px"
-        borderColor="border.overlay"
-      >
-        <Box
-          display="flex"
-          w="full"
-          alignItems="center"
-          justifyContent="space-between"
-        >
-          <Box>
-            <Text
-              fontFamily="heading"
-              fontSize="title.md"
-              fontStyle="normal"
-              fontWeight="bold"
-              lineHeight="24px"
-            >
-              {t("Country")}
-            </Text>
-          </Box>
-          <Box display="flex" gap="16px" alignItems="baseline">
-            <HStack spacing={6} align="start">
-              <FormControl isInvalid={!!errors.cityPopulation}>
-                <FormattedThousandsNumberInput<Inputs>
-                  name="countryPopulation"
-                  control={control}
-                  rules={{
-                    required: t("population-required"),
-                  }}
-                  placeholder={t("country-population-placeholder")}
-                  size="lg"
-                  shadow="1dp"
-                  w="400px"
-                  fontSize="body.lg"
-                  letterSpacing="wide"
-                />
-                <Box display="flex" gap="6px" alignItems="center" py="8px">
-                  <InfoOutlineIcon color="interactive.control" />
-                  <Text
-                    color="content.tertiary"
-                    fontSize="body.md"
-                    letterSpacing="wide"
-                    lineHeight="20px"
-                  >
-                    {t("source")}: {countryPopulationSourceName}
-                  </Text>
-                </Box>
-                <FormErrorMessage
-                  color="content.tertiary"
-                  letterSpacing="0.5px"
-                >
-                  <FormErrorIcon />
-                  {errors.cityPopulation && errors.cityPopulation.message}
-                </FormErrorMessage>
-              </FormControl>
-            </HStack>
-            <InputGroup>
-              <Select
-                placeholder={t("inventory-year-placeholder")}
-                size="lg"
-                w="217px"
-                shadow="1dp"
-                fontSize="body.lg"
-                fontStyle="normal"
-                letterSpacing="wide"
-                _placeholder={{ color: "content.tertiary" }}
-                py="16px"
-                px={0}
-                {...register("countryPopulationYear", {
-                  required: t("inventory-year-required"),
-                })}
-              >
-                {years.map((year: number, i: number) => (
-                  <option value={year} key={i}>
-                    {year}
-                  </option>
-                ))}
-              </Select>
-              <InputRightElement
-                display="flex"
-                alignItems="center"
-                mt={5}
-                mr={6}
-              >
-                {!!year && <CheckIcon color="semantic.success" boxSize={4} />}
-              </InputRightElement>
-            </InputGroup>
-          </Box>
-        </Box>
-      </Box>
-      <Box
-        w="full"
-        py="36px"
-        borderBottomWidth="2px"
-        borderColor="border.overlay"
-      >
-        <Box
-          display="flex"
-          w="full"
-          alignItems="center"
-          justifyContent="space-between"
-        >
-          <Box>
-            <Text
-              fontFamily="heading"
-              fontSize="title.md"
-              fontStyle="normal"
-              fontWeight="bold"
-              lineHeight="24px"
-            >
-              {t("region-or-province")}
-            </Text>
-          </Box>
-          <Box display="flex" gap="16px" alignItems="baseline">
-            <HStack spacing={6} align="start">
-              <FormControl isInvalid={!!errors.cityPopulation}>
-                <FormattedThousandsNumberInput<Inputs>
-                  name="regionPopulation"
-                  control={control}
-                  rules={{
-                    required: t("population-required"),
-                  }}
-                  placeholder={t("region-or-province-population-placeholder")}
-                  size="lg"
-                  shadow="1dp"
-                  w="400px"
-                  fontSize="body.lg"
-                  letterSpacing="wide"
-                />
-                <FormErrorMessage
-                  color="content.tertiary"
-                  letterSpacing="0.5px"
-                >
-                  <FormErrorIcon />
-                  {errors.cityPopulation && errors.cityPopulation.message}
-                </FormErrorMessage>
-              </FormControl>
-            </HStack>
-            <InputGroup>
-              <Select
-                placeholder={t("inventory-year-placeholder")}
-                size="lg"
-                w="217px"
-                shadow="1dp"
-                fontSize="body.lg"
-                fontStyle="normal"
-                letterSpacing="wide"
-                _placeholder={{ color: "content.tertiary" }}
-                py="16px"
-                px={0}
-                {...register("regionPopulationYear", {
-                  required: t("inventory-year-required"),
-                })}
-              >
-                {years.map((year: number, i: number) => (
-                  <option value={year} key={i}>
-                    {year}
-                  </option>
-                ))}
-              </Select>
-              <InputRightElement mt={5} mr={6}>
-                {!!year && <CheckIcon color="semantic.success" boxSize={4} />}
-              </InputRightElement>
-            </InputGroup>
-          </Box>
-        </Box>
-      </Box>
-      <Box
-        w="full"
-        py="36px"
-        borderBottomWidth="2px"
-        borderColor="border.overlay"
-      >
-        <Box
-          display="flex"
-          w="full"
-          alignItems="center"
-          justifyContent="space-between"
-        >
-          <Box>
-            <Text
-              fontFamily="heading"
-              fontSize="title.md"
-              fontStyle="normal"
-              fontWeight="bold"
-              lineHeight="24px"
-            >
-              {t("city")}
-            </Text>
-          </Box>
-          <Box display="flex" gap="16px" alignItems="baseline">
-            <HStack spacing={6} align="start">
-              <FormControl isInvalid={!!errors.cityPopulation}>
-                <FormattedThousandsNumberInput<Inputs>
-                  name="cityPopulation"
-                  control={control}
-                  rules={{
-                    required: t("population-required"),
-                  }}
-                  placeholder={t("city-population-placeholder")}
-                  size="lg"
-                  shadow="1dp"
-                  w="400px"
-                  fontSize="body.lg"
-                  letterSpacing="wide"
-                />
-                <FormErrorMessage
-                  color="content.tertiary"
-                  letterSpacing="0.5px"
-                >
-                  <FormErrorIcon />
-                  {errors.cityPopulation && errors.cityPopulation.message}
-                </FormErrorMessage>
-              </FormControl>
-            </HStack>
-            <InputGroup>
-              <Select
-                placeholder={t("inventory-year-placeholder")}
-                size="lg"
-                w="217px"
-                shadow="1dp"
-                fontSize="body.lg"
-                fontStyle="normal"
-                letterSpacing="wide"
-                _placeholder={{ color: "content.tertiary" }}
-                py="16px"
-                px={0}
-                {...register("cityPopulationYear", {
-                  required: t("inventory-year-required"),
-                })}
-              >
-                {years.map((year: number, i: number) => (
-                  <option value={year} key={i}>
-                    {year}
-                  </option>
-                ))}
-              </Select>
-              <InputRightElement mt={5} mr={6}>
-                {!!year && <CheckIcon color="semantic.success" boxSize={4} />}
-              </InputRightElement>
-            </InputGroup>
-          </Box>
-        </Box>
       </Box>
     </Box>
   );
